open-vocabulary_detection/requirements.txt
transformers
accelerate
huggingface-hub
safetensors

torchvision
pillow
numpy
matplotlib
pyyaml
opencv-python